moved validator query
